+++ du_an_1/Admin/Views/list-lesson.php
@@ -69,9 +69,16 @@
                                                         <?=$course['course_name']?>
                                                     </td>
                                                     <td class="text-center">
-                                                        <a href="index.php?act=deleteLesson&lessonId=<?=$lesson['lesson_id']?>" class="btn btn-primary btn-sm trash deleteItem" type="button" title="Xóa">
-                                                            <i class="fas fa-trash-alt"></i>
-                                                        </a>
+                                                        <?php
+                                                            if ($_SESSION['admin']['role'] == 1) {
+                                                                ?>
+                                                                    <a href="index.php?act=deleteLesson&lessonId=<?=$lesson['lesson_id']?>" class="btn btn-primary btn-sm trash deleteItem" type="button" title="Xóa">
+                                                                        <i class="fas fa-trash-alt"></i>
+                                                                    </a>
+                                                                <?php
+                                                            }
+                                                        ?>
+                                                        
                                                         <?php 
                                                             if ($lesson['lesson_status'] == 1) {
                                                                 ?>
@@ -83,11 +90,12 @@
                                                                 <?php
                                                             }
                                                         ?>
+                                                        
+                                                        
                                                         <a href="index.php?act=editLesson&lessonId=<?=$lesson['lesson_id']?>" class="btn btn-primary btn-sm edit" type="button" title="Sửa" id="show-emp" data-toggle="modal" data-target="#ModalUP">
                                                             <i class="fas fa-edit"></i>
                                                         </a>
-
-                                                        
+                                                                
                                                     </td>
                                                 </tr>
                                             <?php
